+++ yeti/src/test/java/org/mpilone/yeti/StompFrameDecoderTestCase.java
@@ -17,6 +17,120 @@ import io.netty.channel.embedded.EmbeddedChannel;
 public class StompFrameDecoderTestCase {
 
   /**
+   * Tests decoding a frame that is too long because of the header length.
+   */
+  @Test
+  public void testDecodeFrame_LongHeaders() {
+
+    int maxFrameSize = 2 * 1024;
+
+    String header1 = "header1:";
+    for (int i = 0; i < maxFrameSize / 2; ++i) {
+      header1 += 'a';
+    }
+    String header2 = "header2:";
+    for (int i = 0; i < maxFrameSize / 2; ++i) {
+      header1 += 'b';
+    }
+
+    EmbeddedChannel ec =
+        new EmbeddedChannel(new StompFrameDecoder(maxFrameSize));
+
+    final byte[] body = "This is the body.".getBytes(UTF_8);
+
+    ByteBuf buf = Unpooled.buffer();
+    buf.writeBytes(Command.CONNECT.name().getBytes(UTF_8));
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeBytes(header1.getBytes(UTF_8));
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeBytes(header2.getBytes(UTF_8));
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeBytes(body);
+    buf.writeByte(NULL_CHAR);
+    ec.writeInbound(buf);
+
+    Object actual = ec.readInbound();
+    assertNotNull(actual);
+    assertTrue(actual instanceof Frame);
+
+    Frame actualFrame = (Frame) actual;
+    assertTrue(actualFrame.getHeaders().getHeaderNames().contains(
+        StompFrameDecoder.HEADER_BAD_REQUEST));
+  }
+
+  /**
+   * Tests decoding a frame that is too long because of the body length.
+   */
+  @Test
+  public void testDecodeFrame_LongBody() {
+
+    int maxFrameSize = 2 * 1024;
+
+    String body = "";
+    for (int i = 0; i < maxFrameSize; ++i) {
+      body += 'a';
+    }
+
+    EmbeddedChannel ec =
+        new EmbeddedChannel(new StompFrameDecoder(maxFrameSize));
+
+    ByteBuf buf = Unpooled.buffer();
+    buf.writeBytes(Command.CONNECT.name().getBytes(UTF_8));
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeBytes(body.getBytes(UTF_8));
+    buf.writeByte(NULL_CHAR);
+    ec.writeInbound(buf);
+
+    Object actual = ec.readInbound();
+    assertNotNull(actual);
+    assertTrue(actual instanceof Frame);
+
+    Frame actualFrame = (Frame) actual;
+    assertTrue(actualFrame.getHeaders().getHeaderNames().contains(
+        StompFrameDecoder.HEADER_BAD_REQUEST));
+  }
+
+  /**
+   * Tests decoding a frame that is too long because of the body length
+   * indicated in the content length.
+   */
+  @Test
+  public void testDecodeFrame_LongContentLength() {
+
+    int maxFrameSize = 2 * 1024;
+
+    String contentLengthHeader = Headers.CONTENT_LENGTH + ":" + maxFrameSize;
+
+    String body = "";
+    for (int i = 0; i < maxFrameSize; ++i) {
+      body += 'a';
+    }
+
+    EmbeddedChannel ec =
+        new EmbeddedChannel(new StompFrameDecoder(maxFrameSize));
+
+    ByteBuf buf = Unpooled.buffer();
+    buf.writeBytes(Command.CONNECT.name().getBytes(UTF_8));
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeBytes(contentLengthHeader.getBytes(UTF_8));
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeByte(LINE_FEED_CHAR);
+    buf.writeBytes(body.getBytes(UTF_8));
+    buf.writeByte(NULL_CHAR);
+    ec.writeInbound(buf);
+
+    Object actual = ec.readInbound();
+    assertNotNull(actual);
+    assertTrue(actual instanceof Frame);
+
+    Frame actualFrame = (Frame) actual;
+    assertTrue(actualFrame.getHeaders().getHeaderNames().contains(
+        StompFrameDecoder.HEADER_BAD_REQUEST));
+  }
+
+  /**
    * Tests decoding a single, simple frame.
    */
   @Test